Fix materialize js controller
