+++ app/javascript/controllers/index.js
@@ -1,8 +1,7 @@
-import { application } from "controllers/application"
+import { application } from 'controllers/application';
 
-import PostController from "controllers/post_controller"
-application.register("post", PostController)
+import PostController from 'controllers/post_controller';
+application.register('post', PostController);
 
-import initMaterialize from "controllers/materialize"
-
-initMaterialize();
+import MaterializeController from 'controllers/materialize_controller';
+application.register('materialize', MaterializeController);
