+++ app/views/layouts/application.html.erb
@@ -16,7 +16,7 @@
     <%= render partial: 'shared/navbar' %>
     <%= render partial: 'shared/notifications' %>
 
-    <div class="container">
+    <div class="container" data-controller="materialize">
       <%= yield %>
     </div>
